\name{validation.nor}
\alias{validation.nor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Validates the marginal specification of the normal part
}
\description{
This function checks whether mean and variance parameters for the normal part are valid and consistent.
}
\usage{
validation.nor(no.nor, mean.vec.nor = NULL, var.nor = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{no.nor}{
Number of normal variables

}
\item{mean.vec.nor}{
Vector of means for normal variables
}
  \item{var.nor}{
Vector of variances for normal variables
}
}
